+++ CHANGELOG.md
@@ -4,7 +4,19 @@ This changelog follows the specifications detailed in: [Keep a Changelog](https:
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html), although we have not yet reached a `1.0.0` release.
 
 
-## Version 0.10.1 - Unreleased
+## Version 0.10.2 - Unreleased
+
+### Added
+* Add classmethod `PolygonList.random`
+* Add method `PolygonList.to_boxes`
+
+### Fixed
+* `kwimage.Polygon.draw` facecolor argument now properly handles coercible colors.
+* `kwimage.PolygonList.draw` now handles setlim correctly
+* `kwimage.Boxes.coerce` now handles the format argument correctly.
+
+
+## Version 0.10.1 - Released 2024-07-23
 
 ### Added:
 * Added `canonical` flag to Boxes to speed up creation time when input is known to be good.